project v1

+++ pipeline.py
@@ -5,12 +5,25 @@ import re
 
 import requests
 import pandas as pd
+import concurrent.futures
 from loguru import logger
 
 import config
 from models.site_maps import SiteMapsDigikala
 
 sitemap_url = "https://www.digikala.com/sitemap.xml"
+
+
+def process_sitemap_file(extracted_file_path):
+    try:
+        sitemaps_df = pd.read_xml(extracted_file_path)
+        logger.info(f"saved to dataframe: {extracted_file_path}")
+        return sitemaps_df
+    except Exception as e:
+        logger.error(f"Failed to process file: {extracted_file_path}. Error: {e}")
+        return None
+
+
 
 def download_and_extract_gz(url, destination_folder):
     try:
@@ -19,17 +32,22 @@ def download_and_extract_gz(url, destination_folder):
 
         filename = url.split("/")[-1]
         filepath = os.path.join(destination_folder, filename)
+        extracted_file_path = filepath.replace('.gz', '')
 
-        # Check if the file already exists in the destination folder
+        # Check if the extracted file already exists in the destination folder
+        if os.path.exists(extracted_file_path):
+            logger.info(f"Skipping download and extraction. File already exists: {extracted_file_path}")
+            return extracted_file_path
+
+        # Check if the .gz file already exists in the destination folder
         if os.path.exists(filepath):
             logger.info(f"Skipping download. File already exists: {filename}")
-            return filepath
+            return extracted_file_path
 
         with open(filepath, 'wb') as f:
             f.write(gz_url_response.content)
 
         with gzip.open(filepath, 'rb') as f_in:
-            extracted_file_path = filepath.replace('.gz', '')
             with open(extracted_file_path, 'wb') as f_out:
                 # Filter out non-printable characters and write the cleaned content
                 cleaned_content = ''.join(filter(lambda x: x in string.printable, f_in.read().decode()))
@@ -50,39 +68,27 @@ def download_and_extract_gz(url, destination_folder):
         logger.error(f"Failed to download: {url}. Error: {e}")
         return None
 
-def read_sitemap_and_save_to_db(sitemap_file):
+
+def read_sitemap_and_save_to_db(dataframe_batch):
     try:
-        sitemaps_df = pd.read_xml(sitemap_file)
-
-        # Convert NaN values in 'image' column to None to match the database schema
-
-        # Iterate through the DataFrame and save each row to the database
-        for _, row in sitemaps_df.iterrows():
-            # Check if the 'loc' value already exists in the database
-            existing_loc_record = SiteMapsDigikala.select().where(SiteMapsDigikala.loc == row['loc']).first()
-            if not existing_loc_record:
-                SiteMapsDigikala.create(
-                    loc=row['loc'],
-                    changefreq=row['changefreq'],
-                    priority=row['priority'],
-                )
-            else:
-                logger.info(f"Skipping duplicate record: {row['loc']}")
-
-        logger.info(f"Data from {sitemap_file} saved to the database.")
+        dataframe_batch = dataframe_batch.drop(columns=['image'], errors='ignore')
+        records = dataframe_batch.to_dict(orient='records')
+        num_records = len(records)
+        SiteMapsDigikala.insert_many(records).execute()
+        logger.info(f"Saved {num_records} records to the database.")
     except Exception as e:
-        logger.error(f"Failed to process {sitemap_file}. Error: {e}")
+        logger.error(f"Failed to save data to the database. Error: {e}")
+
 
 def main():
     try:
         get_site_map = requests.get(sitemap_url)
-        get_site_map.raise_for_status()  
+        get_site_map.raise_for_status()
         sitemap_xml = get_site_map.text
     except requests.exceptions.RequestException as e:
         logger.error(f"Failed to download sitemap.xml. Error: {e}")
         sitemap_xml = ""
 
-    # Find all the .gz URLs in the sitemap.xml
     gz_urls = re.findall(r'<loc>(.*?)</loc>', sitemap_xml)
 
     # Destination folder to store the extracted files
@@ -97,12 +103,28 @@ def main():
             extracted_file_paths.append(extracted_file_path)
             logger.info(f"Extracted: {extracted_file_path}")
 
+    # Create an empty DataFrame to store the combined sitemaps
+    combined_sitemaps_df = pd.DataFrame()
+
+
     # Read the contents of each sitemap file into a Pandas DataFrame
-    for i, extracted_file_path in enumerate(extracted_file_paths):
-        try:
-            read_sitemap_and_save_to_db(extracted_file_path)
-        except Exception as e:
-            logger.error(f"Failed to process DataFrame {i}. Error: {e}")
+    with concurrent.futures.ProcessPoolExecutor() as executor:
+        # Submit tasks to read and process sitemap files
+        futures = [executor.submit(process_sitemap_file, extracted_file_path) for extracted_file_path in extracted_file_paths]
+
+        # Combine the results into a single DataFrame
+        combined_sitemaps_df = pd.concat([future.result() for future in concurrent.futures.as_completed(futures)], ignore_index=True)
+
+    # Remove duplicates based on 'loc' column
+    combined_sitemaps_df.drop_duplicates(subset='loc', inplace=True)
+
+
+    # Save data in batches of 10,000 records
+    batch_size = 10000
+    for batch_start in range(0, len(combined_sitemaps_df), batch_size):
+        dataframe_batch = combined_sitemaps_df.iloc[batch_start:batch_start + batch_size]
+        read_sitemap_and_save_to_db(dataframe_batch)
+
 
 if __name__ == "__main__":
     main()
